Add tailwind error handling

+++ app/main.py
@@ -1,6 +1,7 @@
 from contextlib import asynccontextmanager
 from subprocess import Popen, PIPE
 from typing import Optional
+import os
 
 from fastapi import FastAPI, Header, Request, Response, status
 from fastapi.responses import HTMLResponse
@@ -14,11 +15,21 @@ from app.routers import auth_routes, login_routes, user_routes
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     if ENV_MODE == "DEV":
-        process = Popen(
-            ["./tailwindcss", "-o", "static/tailwind.css"], stdout=PIPE, stderr=PIPE
-        )
-        yield
-        process.terminate()
+        process = None
+        try:
+            if not os.path.isfile("./tailwindcss"):
+                raise FileNotFoundError("tailwindcss command not found")
+            process = Popen(
+                ["./tailwindcss", "-o", "static/tailwind.css"], stdout=PIPE, stderr=PIPE
+            )
+            yield
+        except FileNotFoundError as e:
+            print(f"Error: {e}. TailwindCSS process not started.")
+            yield
+        finally:
+            if process:
+                process.terminate()
+                process.wait()
     else:
         yield
